Adjust default HTTP response code colors for xymonnet
git-svn-id: svn://svn.code.sf.net/p/xymon/code@7698 44351d6e-118b-4698-b696-ce33095ecaa4

+++ branches/4.x-master/xymonnet/contest.h
@@ -171,7 +171,7 @@ typedef struct {
 	unsigned char	*output;		/* Data from server */
 	unsigned int	outlen;
 
-	long		httpstatus;		/* HTTP status from server */
+	int		httpstatus;		/* HTTP status from server */
 	char		*contenttype;		/* Content-type: header from server */
 
 	int		contentcheck;		/* 0=no content check, 1=regex check, 2=digest check */
